Updated to v3.0.0 ALPHA
Added in the ability to see the number of total attempts for LoS and Distance.
Added the ability to change from plugin spawning to Valve spawning through a cvar, this isn't changable on the fly, the map has to be reloaded.
Added a cvar to set the amount of distance attempts along with LoS attempts, changed the default value of 10 to 64.